requirements.txt

elevenlabs
python-dotenv
pydub
langchain
ffmpeg-python
langchain-community
fastapi
uvicorn
pydantic
